refactor(RacingCar) : TrialCount를 만들어서 시도 횟수 검증 메서드 분리
