+++ src/main/java/racingcar/controller/RacingController.java
@@ -7,13 +7,14 @@ import static racingcar.view.OutputView.printResultMessage;
 import static racingcar.view.OutputView.printTurnResult;
 
 import racingcar.domain.Cars;
+import racingcar.domain.TrialCount;
 
 public class RacingController {
 
     public void play() {
         try {
             Cars cars = new Cars(insertName());
-            int trialCount = insertNumber();
+            TrialCount trialCount = insertNumber();
 
             progressTurns(cars, trialCount);
             printMessage(cars.getWinners());
@@ -22,10 +23,10 @@ public class RacingController {
         }
     }
 
-    private void progressTurns(Cars cars, int trialCount) {
+    private void progressTurns(Cars cars, TrialCount trialCount) {
         printResultMessage();
 
-        for (int i = 0; i < trialCount; i++) {
+        for (int i = 0; i < trialCount.getTrialCount(); i++) {
             cars.moveCars();
             printTurnResult(cars);
         }
