+++ src/main/java/racingcar/domain/TrialCount.java
@@ -1,14 +1,21 @@
-package racingcar.utils;
+package racingcar.domain;
 
-public class Validator {
+public class TrialCount {
 
-    private static final int MIN_LENGTH = 0;
+    private static final int MIN_COUNT = 0;
     private static final String ERROR_NUMBER_OVER_ONE = "[ERROR] 횟수는 1 이상이어야 합니다.";
     private static final String ERROR_NUMBER_NOT_DIGIT = "[ERROR] 횟수는 숫자로 입력해야 합니다.";
 
-    public static int toInt(String text) {
+    private final int trialCount;
+
+    public TrialCount(final String trialCount) {
+        this.trialCount = toInt(trialCount);
+    }
+
+    private int toInt(final String text) {
         try {
             int parseInt = Integer.parseInt(text);
+
             checkPositiveNumber(parseInt);
             return parseInt;
         } catch (NumberFormatException e) {
@@ -16,10 +23,13 @@ public class Validator {
         }
     }
 
-    private static void checkPositiveNumber(int parseInt) {
-        if (parseInt <= MIN_LENGTH) {
+    private void checkPositiveNumber(final int parseInt) {
+        if (parseInt <= MIN_COUNT) {
             throw new IllegalArgumentException(ERROR_NUMBER_OVER_ONE);
         }
     }
 
+    public int getTrialCount() {
+        return trialCount;
+    }
 }
