+++ src/main/java/racingcar/view/InputView.java
@@ -1,9 +1,9 @@
 package racingcar.view;
 
-import static racingcar.utils.Validator.toInt;
 import static racingcar.view.OutputView.printCarNamesMessage;
 import static racingcar.view.OutputView.printNumberMessage;
 
+import racingcar.domain.TrialCount;
 import racingcar.utils.Console;
 
 public class InputView {
@@ -17,9 +17,8 @@ public class InputView {
         return input.split(SPLIT_DELIMITER, SPLIT_LIMIT);
     }
 
-    public static int insertNumber() {
+    public static TrialCount insertNumber() {
         printNumberMessage();
-        String input = Console.readLine();
-        return toInt(input);
+        return new TrialCount(Console.readLine());
     }
 }
